Check for http:// prefix in MasterAddress.
When the client uses the env. variable ROS_MASTER_URI for the master address definition
in node creation, then a conflict between roscore and the node appears.
The problem is that roscore accepts ROS_MASTER_URI with the prefix http:// in contrast
to the Go net.ResolveTCPAddr method that requires to omit that prefix.
This fix checks conf.MasterAddress for prefix presence and trims it if necessary.
Finally, it will be possible to use the env. variable ROS_MASTER_URI both with goroslib and roscore.