fix(xmlrpc): workaround for non-standard encoding
Certain clients include `encoding=utf8` in the xmlrpc message which
causes an error about a missing charset reader. This fix checks that
if we get an "utf8" encoding we just proxy it through. If other charsets
are received we error out (same as it would without this fix).

This all comes from the fact that the xml decoder checks for "utf-8"
charset, not "utf8".

The offending client where I found this is the Foxglove studio.